+++ HISTORY.rst
@@ -98,6 +98,7 @@ Unreleased
 - fix: #288 add_table accepts float dimensions
 - fix: #608 Name of placeholder in layout gets reset to placeholder basename when adding slide
 - #68 feature: reorder a slide
+- #67 feature: delete a slide
 
 1.0.2 (2024-08-07)
 ++++++++++++++++++